Moved stories to be under stories folder
