+++ component-library/src/stories/Checkbox.stories.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Checkbox from './Checkbox';
+import Checkbox from '../components/Checkbox/Checkbox';
 import { withDesign } from 'storybook-addon-designs'
 
 export default {
