+++ component-library/src/stories/Input.stories.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import Input from './Input';
+import Input from '../components/Input/Input';
 import { withDesign } from 'storybook-addon-designs'
 
 export default {
