+++ component-library/src/stories/Tooltip.stories.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { Tooltip } from './Tooltip';
+import { Tooltip } from '../components/Tooltip/Tooltip';
 import { withDesign } from 'storybook-addon-designs'
 
 export default {
